id must be i32

+++ src/database/mod.rs
@@ -542,7 +542,7 @@ impl DataBaseConnection {
 
     pub  fn update_region(&mut self, region: &Region) -> bool {
         match self.postgres.execute("UPDATE regions SET name=$1, transport_company=$2, frequency=$3, protocol=$4 WHERE id=$5",
-                              &[&region.name, &region.transport_company, &(region.frequency as i64), &region.protocol, &(region.id as i64)]) {
+                              &[&region.name, &region.transport_company, &(region.frequency as i64), &region.protocol, &(region.id as i32)]) {
             Err(e) => {
                 println!("Error updating region: {:?}", e);
                 false
